feat(PCS-6): PgVendorCallService invokeVendor + applyOutcome TX 경계 분리
TX 없음 구간에서 벤더 HTTP 대기, TX_B 안에서만 RDB 변경하는 구조로 분리.

- GatewayOutcome sealed interface 별도 파일 추출 (package-private)
- invokeVendor (TX 없음): 벤더 HTTP 호출 후 GatewayOutcome 반환
- applyOutcome (@Transactional TX_B): 5분기 처리
  - Success → APPROVED 종결 + Outbox INSERT
  - NonRetryable → FAILED 종결 + Outbox INSERT
  - Retryable 잔여 시도 → 재시도 명령 INSERT (IN_PROGRESS 유지)
  - Retryable 소진 → DLQ 명령 INSERT (IN_PROGRESS 유지)
  - HandledInternally → DuplicateApprovalHandler 위임
- callVendor @Deprecated(forRemoval=true) 유지 (PCS-9 호출처 교체 예정)
- DuplicateApprovalHandler 생성자 주입 추가
- PaymentConfirmConsumerTest / PgVendorCallServiceVendorTypeTest 컴파일 오류 수정
- ./gradlew test 238 PASS / 0 FAIL
